componentised errortext, hinttext, label and labeltext to atomic scale
