+++ src/components/inputField/index.js
@@ -6,97 +6,30 @@ import PropTypes from "prop-types";
 import glamorous from "glamorous";
 
 import * as COLOUR from "govuk-colours";
-import { BREAKPOINTS, SITE_WIDTH } from "../../constants/index";
+import { FONT_SIZE, MEDIA_QUERIES, NTA_LIGHT } from "../../constants/index";
 
-const mediaQueries = {
-  largeScreen: `@media only screen and (min-width: ${BREAKPOINTS.LARGESCREEN})`
-};
-
-const Label = glamorous.label(
-  {
-    marginBottom: "15px",
-    display: "flex",
-    flexDirection: "column",
-    boxSizing: "border-box",
-    [mediaQueries.largeScreen]: {
-      maxWidth: SITE_WIDTH,
-      marginBottom: "30px"
-    }
-  },
-  ({ errorText }) => ({
-    borderLeft: errorText ? `4px solid ${COLOUR.ERROR_COLOUR}` : "",
-    marginRight: errorText ? `15px` : "0",
-    paddingLeft: errorText ? `10px` : "0"
-  })
-);
-
-const LabelText = glamorous.span(
-  {
-    fontFamily: '"nta", Arial, sans-serif',
-    WebkitFontSmoothing: "antialiased",
-    MozOsxFontSmoothing: "grayscale",
-    fontWeight: 400,
-    fontSize: "16px",
-    lineHeight: "1.25",
-    color: `${COLOUR.BLACK}`,
-    paddingBottom: "2px",
-    [mediaQueries.largeScreen]: {
-      fontSize: "19px",
-      lineHeight: "1.31579"
-    }
-  },
-  ({ errorText }) => ({
-    fontWeight: errorText ? 700 : LabelText.fontWeight
-  })
-);
-
-const HintText = glamorous.span({
-  fontFamily: '"nta", Arial, sans-serif',
-  WebkitFontSmoothing: "antialiased",
-  MozOsxFontSmoothing: "grayscale",
-  fontWeight: 400,
-  textTransform: "none",
-  fontSize: "16px",
-  lineHeight: "1.25",
-  color: `${COLOUR.GREY_1}`,
-  [mediaQueries.largeScreen]: {
-    fontSize: "19px",
-    lineHeight: "1.31579"
-  }
-});
-
-const ErrorText = glamorous.span({
-  fontFamily: '"nta", Arial, sans-serif',
-  WebkitFontSmoothing: "antialiased",
-  MozOsxFontSmoothing: "grayscale",
-  fontWeight: 700,
-  textTransform: "none",
-  fontSize: "16px",
-  lineHeight: "1.25",
-  color: `${COLOUR.ERROR_COLOUR}`,
-  paddingTop: "4px",
-  paddingBottom: "2px",
-  [mediaQueries.largeScreen]: {
-    fontSize: "19px",
-    lineHeight: "1.31579"
-  }
-});
+import LabelText from "../labelText/index";
+import ErrorText from "../errorText/index";
+import HintText from "../hintText/index";
+import Label from "../label/index";
 
 const Input = glamorous.input(
   {
     boxSizing: "border-box",
-    fontFamily: '"nta", Arial, sans-serif',
+    fontFamily: NTA_LIGHT,
     WebkitFontSmoothing: "antialiased",
     MozOsxFontSmoothing: "grayscale",
     fontWeight: 400,
     textTransform: "none",
-    fontSize: "16px",
-    lineHeight: "1.25",
+    fontSize: FONT_SIZE.SIZE_16,
+    lineHeight: "20px",
     width: "100%",
     padding: "5px 4px 4px",
     border: `2px solid ${COLOUR.BLACK}`,
-    [mediaQueries.largeScreen]: {
-      width: "50%"
+    [MEDIA_QUERIES.LARGESCREEN]: {
+      width: "50%",
+      fontSize: FONT_SIZE.SIZE_19,
+      lineHeight: "25px"
     },
     "[disabled]": {
       cursor: "auto"
